Improved functionality for expanding texts
* The show all-link has been moved to the center of each story and this
has been implemented in all views.
* The link texts have been changed into small images of arrows. These
arrows have been placed in the middle of each story element.
* In order to make it look good when going into edit mode, I also had to
make a new fix for the bouncing <li>-bug. It was instead solved by
setting a fixed width of each <li>-element and updating it when the
window size changes.

Left to do in this area is:
* Possibly change the design with arrow buttons.
* Possibly remember which descriptions that were extended before going
into edit mode since all stories are currently collapsed again when
exiting edit mode again.
* Possibly collapse all extended stories when sorting begins.